[native] Migrate ActionResultModal component to Reanimated v2 API
Summary:
A straightforward migration: use progressV2 shared value and use useAnimatedStyle

Depends on D14129

Test Plan:
Flow and I'd like to navigate to this modal in the app but as far as I searched there is no way to navigate to it so I'm wondering if this modal is still in use. If it's no longer used we can just remove it?

Reviewers:
ashoat, tomek

Subscribers: