[native] Move onPress and onLayout methods in MultimediaMessageMultimedia to functional component
Summary:
Move onPress and onLayout methods to functional component as a part of MultimediaMessageMultimedia migration to a functional component. Empty onLayout looks weird but in the history I can see it was introduced to fix some Android measure bug and I believe it so I'm leaving it as it was.

Depends on D14131

Test Plan: Open up a multimedia message and see if it behaves correctly.

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14132