[native] Migrate opacity in MultimediaMessageMultimedia to Reanimated v2 API in functional component
Summary:
Originally opacity was held in a state because the component needed to be rerendered every time it changed from a number (if there is no overlay position) to a reanimated node (or the other way). And we used opacity only in animatedWrapperStyle.  But in reanimated v2 we can assign a number or interpolated value in useAnimatedStyle so we don’t need a state anymore.
Because migrating the state to a functional component and then getting rid of it anyway when migrating to reanimated v2 api  seems like unnecessary hassle, I migrated it in one diff

Depends on D14132

Test Plan: Open up a multimedia message and see if it behaves correctly.

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14133