[native] Move componentDidUpdate logic in MultimediaMessageMultimedia to functional component
Summary:
Move the logic in componentDidUpdate to a functional component, that is: if the scroll in overlay context is now disabled and it wasn't disabled before, then call setClickable(true). Or in more human way: if a modal that was previously blocking scroll etc. disappeared then the message with the multimedia is clickable again.

Depends on D14133

Test Plan:
Open a modal (a tooltip or click a message with an image)
Close a modal
Click on a message with an image
Verify the image is expanded

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14134