[native] implement QR Auth context
Summary:
[ENG-9959](https://linear.app/comm/issue/ENG-9959/improve-scanning-qr-code-flow).

We need to wrap this logic in context so this will be available on all the screens in the flow.

This diff only introduces the context, not changing the logic - just moving it to the context.

In [ENG-9968](https://linear.app/comm/issue/ENG-9968/final-testing-task) there is an example of how this looks like

Depends on D14101

Test Plan: Test adding secondary device - should work exactly the same.

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14105