[native] implement screen when scanning QR code a non-primary device
Summary:
[ENG-9959](https://linear.app/comm/issue/ENG-9959/improve-scanning-qr-code-flow).

Copy is taken from [ENG-9936](https://linear.app/comm/issue/ENG-9936/show-a-specific-error-when-scanning-a-qr-code-on-a-secondary-device)

Styles are taken from similar register screens, and there is a chance that those will duplicate with other QR Auth screens - if so I am going to refactor this but for now I prefer to keep this like this if there are any changes needed specific to a screen.

Also, after this is done maybe worth renaming the `<Registration*` components to `<Auth*` components to avoid confusion.

In [ENG-9968](https://linear.app/comm/issue/ENG-9968/final-testing-task) there is an example of what the flow looks like.

{F3446213}

Depends on D14105

Test Plan: Navigate to screen

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14114