[native] implement screen when connecting secondary device
Summary:
[ENG-9959](https://linear.app/comm/issue/ENG-9959/improve-scanning-qr-code-flow).

Copy is taken from existing Alert.

This diff is just UI, navigating to it and logic is implemented later in the stack.

In [ENG-9968](https://linear.app/comm/issue/ENG-9968/final-testing-task) there is an example of how the flow looks like.

{F3446216}

Depends on D14114

Test Plan: Navigate to screen

Reviewers: ashoat, tomek

Reviewed By: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14115