[native] Stop resetting ChatThreadList on navigation
Summary:
Currently, if the user scrolls down to a chat and then opens it, and then goes back, the `ChatThreadList` will not appear the same as before, and their chat may not be visible.

I think this is confusing behavior in the first place. It gets especially bad when the `ChatThreadList` is reloading its order async.

Depends on D14143

Test Plan: I tested this task by playing around with the `ChatThreadList` on mobile while using a stopwatch to measure how long various operations took. I tested the updated code 3 times against both the most recent Testflight build as well as `master`. I found that performance was approximately the same as before. I tested scrolling down, selecting threads that were read, selecting threads that were unread, and search.

Reviewers: tomek, angelika

Differential Revision: https://phab.comm.dev/D14144