[native] implement screen after secondary device is connected
Summary:
[ENG-9959](https://linear.app/comm/issue/ENG-9959/improve-scanning-qr-code-flow).

Final screen when adding secondary device succeded. There is no option to go back, because it could be strange for the user to back on connecting screen.

In [ENG-9968](https://linear.app/comm/issue/ENG-9968/final-testing-task) there is example of how this looks like

Depends on D14116

Test Plan:
Navigate to screen

Reviewers: ashoat, tomek