[native] implement screen after secondary device is connected
Summary:
[ENG-9959](https://linear.app/comm/issue/ENG-9959/improve-scanning-qr-code-flow).

The final screen when adding a secondary device succeeded. There is no option to go back because it could be strange for the user to back to the connecting screen.

In [ENG-9968](https://linear.app/comm/issue/ENG-9968/final-testing-task) there is an example of how this looks like

{F3481879}

Depends on D14116

Test Plan: Navigate to screen

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14149