[native] add danger option to `PrimaryButton`
Summary:
[ENG-9959](https://linear.app/comm/issue/ENG-9959/improve-scanning-qr-code-flow).

According to [this](https://linear.app/comm/issue/ENG-8226/remove-secondary-device-from-device-list-if-primary-doesnt-receive#comment-dea392ea) we want a red `Remove` button so adding this config here.

I was trying to follow the approach we had in the codebase and was looking for patterns, but perhaps I broke some rules in how we should use colors, so sorry for that in advance.

Dark:
{F3449358}

Light:
{F3449360}

Example screen where this is used:
{F3449364}

Depends on D14149

Test Plan: Render button using dark and light theme

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14150