[native] add danger option to `PrimaryButton`
Summary:
[ENG-9959](https://linear.app/comm/issue/ENG-9959/improve-scanning-qr-code-flow).

According to [this](https://linear.app/comm/issue/ENG-8226/remove-secondary-device-from-device-list-if-primary-doesnt-receive#comment-dea392ea) we want red `Remove` button so adding this config here.

I was trying to follow approach we had in the codebase and was looking for pattern, but perhaps I broke some rule in how we should use colors, so sorry for that in advance.

Depends on D14149

Test Plan:
Render button using dark and light theme

Reviewers: ashoat, tomek