[native] add callback to remove secondary device during QR auth
Summary:
[ENG-9959](https://linear.app/comm/issue/ENG-9959/improve-scanning-qr-code-flow).

According to [this](https://linear.app/comm/issue/ENG-8226/remove-secondary-device-from-device-list-if-primary-doesnt-receive#comment-dea392ea) we want to be able to remove device on a timeout.

Depends on D14150

Test Plan: Call this callback when QR auth flow is finished - devices is removed from device list, secondary device is visually logged out.

Reviewers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14151