[native] implement screen when secondary device is not responding
Summary:
[ENG-9959](https://linear.app/comm/issue/ENG-9959/improve-scanning-qr-code-flow).

Copy was discussed in [ENG-8226#comment-dea392ea](https://linear.app/comm/issue/ENG-8226/remove-secondary-device-from-device-list-if-primary-doesnt-receive#comment-dea392ea).

I am not sure how to solve this in terms of UX, the time when removing the device is pending:
1. Like this, there might be a lag before navigating to Linked devices.
2. Not await promise, but in this case user might see the device on the list, and will disappear after this promise is finished.
3. Somehow show progress, but don't know how to handle the `Ignore` button then.

Also, swallowing errors here and just navigating to the Linked devices screen, we could show an Alert but wondering if this is worth it.

In [ENG-9968](https://linear.app/comm/issue/ENG-9968/final-testing-task) there is an example of how this looks like.

{F3449501}

Depends on D14151

Test Plan: Navigate to this screen

Reviewers: ashoat, tomek

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14152