[lib] Use ChatThreadItemLoaderCache for loading sidebar
Summary:
We have the same problem for sidebars that we have at the parent level, especially for the "Daily updates" channel, which has close to 1000 sidebars inside of it.

We need to avoid creating a `Promise` for each sidebar. We can use `ChatThreadItemLoaderCache` to achieve this.

Depends on D14144

Test Plan: I tested this task by playing around with the `ChatThreadList` on mobile while using a stopwatch to measure how long various operations took. I tested the updated code 3 times against both the most recent Testflight build as well as `master`. I found that performance was approximately the same as before. I tested scrolling down, selecting threads that were read, selecting threads that were unread, and search.

Reviewers: tomek, angelika

Subscribers: