[lib] Fix stack overflow from lastUpdatedTime
Summary: This resolves [ENG-10016](https://linear.app/comm/issue/ENG-10016/maximum-call-stack-size-exceeded-after-registration). It looks like when this was changed from an immediately-invoked `Promise` to a function that returns a `Promise`, we ended up in a situation where the `lastUpdatedTime` dereference ended up resolving to the new value rather than the prior value.

Test Plan: Prior to this change, I was able to reproduce the issue in the Linear task. Afterwards, it no longer reproduces

Reviewers: tomek, angelika, varun, will

Reviewed By: will

Differential Revision: https://phab.comm.dev/D14156