[lib] Add log and continue in getPushUserInfo when thread missing from ThreadStore
Summary:
This has come up twice now: [ENG-9454](https://linear.app/comm/issue/ENG-9454/cannot-read-properties-of-undefined-when-leaving-thread) and [ENG-9946](https://linear.app/comm/issue/ENG-9946/cannot-read-properties-of-undefined-reading-members).

I think it will help us in the future to have a more specific error message. It also is probably a good idea to `continue;` here, in case there are other notifs that we can send successfully.

Test Plan: Ran the code and confirmed the log was printed

Reviewers: angelika, kamil, tomek

Subscribers: