[native] Fall back to v1 login when the user doesn't have a backup
Summary:
Check if the user has a backup. When they don't have it, use the v1 login flow.

https://linear.app/comm/issue/ENG-9620/client-restore-flow-fall-back-to-v1-when-missing-backup

Depends on D14165

Test Plan: Because pf https://linear.app/comm/issue/ENG-10039/allow-checking-if-a-user-has-a-backup I had to hack it a bit: instead of calling `retrieveLatestBackupInfo` I hardcoded its return value to null. Then for both password and SIWE user verified that it is possible to log in.

Reviewers: kamil, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14180