[lib] Wait for tunnelbroker and keyserver connection when updating relationships
Summary:
https://linear.app/comm/issue/ENG-9547/dms-created-when-ashoat-opened-web-app-after-not-touching-it-for-a

Implemented solution described by Ashoat here https://linear.app/comm/issue/ENG-9547/dms-created-when-ashoat-opened-web-app-after-not-touching-it-for-a#comment-5c104d82

`waitForConnection` is a bit ambiguous name but I couldn't think of a better one. `waitForTunnelbrokerAndKeyserverConnection` seems to long? I'm open to suggestions

Test Plan:
Ideally it would be better to test it with Farcaster but I don't have Farcaster account so to simulate the issue:
1. Have mobile and web clients that do not know each other
2. Put a sleep() call before connecting to tunnelbroker socket (for example 20s)
3. Close web client
4. Send a DM on mobile to web. Make sure it's sent (checkmark)
5. Open web client
6. Quickly (before sleep from step 2 ends) add mobile to friends (with Settings -> friends -> add friends)
Before the changes the thread was duplicated
After the changes the app waits for tunnelbroker connection and only one thread is created

Reviewers: tomek, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14167