[native] rename `<Registration*` components to `<Auth*` components
Summary:
This is a follow-up to what I suggested in D14114 - to avoid confusion renaming to something more generic as now those components are also used outside of registration flow.

This diff is only rename, I am going to move them to a different directory in next diff to make review easy.

Depends on D14169

Test Plan: Flow

Reviewers: bartek, tomek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14170