[lib] Always include an extra item in getCandidateSidebarItemsForThreadList
Summary: We need the extra item to determine whether to show the "See more..." item. Currently there's a bug where it is omitted when it should be displayed.

Test Plan: I tested in my local environment by creating a bunch of threads and confirming that "See more..." would appear on a web client. Tried from two user accounts, one where it was read and one where it was unread.

Reviewers: tomek, angelika

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D14175