[keyserver] Avoid ignorePromiseRejections if isScriptViewer
Summary:
When we use `ignorePromiseRejections` inside a script, it can cause problems because the SQL connection can be shut down before the operation concludes.

As such, we have a pattern to `await` in those cases. You can see we're already using that pattern in the same file for `deletionUpdatesPromise`.

Test Plan: Relying on Flow here, as this is a pretty clear and established pattern

Reviewers: tomek, angelika

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D14176