[native] Update device list as a part of the restoration
Summary:
After the restoration, the device list in the store is empty, which means that the current device can't be recognized as the primary one. To solve that we can update the store as a part of the success restoration action.

https://linear.app/comm/issue/ENG-10033/regenerate-a-new-backup

Depends on D14180

Test Plan: Add some logs to the code that checks if a device is primary and check that it returns true just after the restoration. Also, check the SIWE restoration - a screen on which a new backup is created should be shown just after the restoration - this verifies that the device was successfully recognized as primary because the screen is only shown on the primary devices.

Reviewers: kamil, bartek

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14182