[native] Merge registration and sign in navigators
Summary:
As a part of the registration flow we might need to open a screen from the restoration navigator and continue with the restoration flow. To make this interaction better, we need to have both flows in the same navigation structure.

https://linear.app/comm/issue/ENG-10040/handle-existing-wallet-registration

Depends on D14186

Test Plan: Check if it is possible to open both registration and restoration screens.

Reviewers: kamil, bartek

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14192