[native] Handle the restore flow as a part of the registration flow
Summary:
When a user has an existing Ethereum account, they can try to use it to register a new account. In the old flow, we would show a screen when they can choose to either log in or use a different wallet - we're keeping this functionality for the v1 users.
In the new flow, there's no way to just log in - it is either restore or secondary login. The screen is updated so that the user can choose to restore their account, open the secondary login flow (QR), or choose a different account.

https://linear.app/comm/issue/ENG-10040/handle-existing-wallet-registration

Depends on D14193

Test Plan:
Tested that both new and old flows work (shown on the videos).

Reviewers: kamil, bartek