[native] Fix SIWE panel size
Summary:
The only purpose of `onSuccessOrCancel` is to stop the view from resizing after we're done with all the work. It was incorrectly called in the effect, because then the work wasn't yet completed. The intention behind calling this function there was to avoid continuing the effect when the nonce is already there - but there's no need to do that because early return already protects us against it.

https://linear.app/comm/issue/ENG-10027/siwe-wallet-connect-ui-is-shifted-when-opened-with-some-request-data

Depends on D14194

Test Plan:
Tested on iOS simulator - both social proof and backup decrypt steps should have the SIWE panel displayed correctly.

Reviewers: kamil, bartek, ashoat