[native] fix not staring backup handler when performing migration to new flow
Summary:
[ENG-10094](https://linear.app/comm/issue/ENG-10094/migration-to-v2-is-not-working-on-non-primary-device).

This is actual fix to the issue, we might want to migrate on a non-primary device. In that case we need to additionally run backup handler.

Depends on D14214

Test Plan: Tested if issue from the ENG-10094 - Identity is called but migration fails because of [ENG-10101](https://linear.app/comm/issue/ENG-10101/non-primary-device-is-unable-to-migrate-to-v2).

Reviewers: bartek, tomek