[native] Migrate gestures to RNGH V2 API in CameraModal
Summary:
https://linear.app/comm/issue/ENG-8152/convert-cameramodal-to-reanimated-2-syntax

I've broken the process of migrating zoom and focus in camera modal into several diffs, but between the diffs zoom and focus is broken.

In this diff I've replaced `PinchGestureHandler` and `TapGestureHandler` with gesture-handler v2 API: `GestureDetector`.
A special care must be taken so that we can detect pinch and tap gestures simultanously, that's why pinch goes first and then tap goes
second in `Gesture.Exclusive`.

Depends on D14230

Test Plan: Verify focus and zoom works in the next diffs.

Reviewers: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14231