[native] Migrate send button style in CameraModal to functional component
Summary:
https://linear.app/comm/issue/ENG-8152/convert-cameramodal-to-reanimated-2-syntax

This animation is in RN's Animated API. I considered changing it to reaniamted api, but `SendMediaButton` is used in several places
and I'd have to change them all. So I just moved the animation code to functional component.

Depends on D14236

Test Plan: Verify send button is animated correctly

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14237