[native] Fix focus animation on devices with top notches in CameraModal
Summary:
https://linear.app/comm/issue/ENG-8152/convert-cameramodal-to-reanimated-2-syntax

When testing the camera modal migration on a device with top notch (iPhone 16) I've noticed that
the focus animation runs when I click the camera buttons. This is not supposed to happen, because we
have `outsideButtons` function that's supposed to check for that. It also happens on the master branch
so it's not introduced with the migration.

It turns out that `outsideButtons` function doesn't account for the top notch.

Depends on D14240

Test Plan:
Verify focus works.
Verify focus animation is not triggered when clicking camera buttons.

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14241