[native][lib] implement `useDeviceKind` hook
Summary:
Part of [ENG-10094](https://linear.app/comm/issue/ENG-10094/migration-to-v2-is-not-working-on-non-primary-device) and [ENG-10103](https://linear.app/comm/issue/ENG-10103/uploading-backup-after-restoreregistration-sometimes-not-work).

This is a more convenient way of using the same logic as in `useCheckIfPrimaryDevice` - it reduces the asynchronous nature of effects and allows for simplifying code. This is also more detailed because it returns an `unknown` value which is different from `!isPrimary` which could be interpreted as being a secondary device.

Using this hook should be safe and synced with the actual state because the primary can change in two scenarios:
1. Device list updated - it is reactive to `userDevicesInfos`
2. `deviceID` change - this can't be done without logging out, and this hook is reactive to `loggedIn`.

In the future, it should also be reactive to primary keys rotation, but this protocol is not yet implemented.

Depends on D14217

Test Plan: Tested on secondary and primary

Reviewers: bartek, tomek

Reviewed By: bartek, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14242