[native][lib] implement `useDeviceKind` hook
Summary:
Part of [ENG-10094](https://linear.app/comm/issue/ENG-10094/migration-to-v2-is-not-working-on-non-primary-device) and [ENG-10103](https://linear.app/comm/issue/ENG-10103/uploading-backup-after-restoreregistration-sometimes-not-work).

This is a more convinient way of using this hook - it reduces the asynchronous nature of effects and allows simplifying code.

Depends on D14217

Test Plan: Tested on secondary and primary

Reviewers: bartek, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14242