[native] extract code in `BackupHandler` to callbacks
Summary:
Part of [ENG-10094](https://linear.app/comm/issue/ENG-10094/migration-to-v2-is-not-working-on-non-primary-device) and [ENG-10103](https://linear.app/comm/issue/ENG-10103/uploading-backup-after-restoreregistration-sometimes-not-work).

Moving code around to simplify complex effects, this adds some code duplication, but here it worth it as on the other hand, code is hard to read and mantain.

Conditions are simplified in the next diff in the stack.

Depends on D14243

Test Plan: Tested later in the stack

Reviewers: bartek, tomek