[native][web][lib] clean `unauthorizedDeviceID` after abandoning secondary device auth
Summary:
This is a fix for [ENG-10093](https://linear.app/comm/issue/ENG-10093/scanning-qr-code-immediately-after-registration-sometimes-fails).

The issue was that when showing the QR code, we created an unauthorized Tunnelbroker connection and it was kept until log in or failure - but we should also disconnect when the screen is dissmissed. Because while navigating to restore QR code is rendered, we had an issue that Tunnelbroker connection was not working until killing and openning the app.

Test Plan:
Check logs about what type of TB connection is created when navigation to/from QR code screen.
Tested adding secondary device right after scanning.

Reviewers: bartek, tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14247