[lib] Update the DMs status to failed after fetching from the DB
Summary:
After the app is restarted, all the messages that weren't sent fully should be marked as failed so that they can be retried. We wanted to do that but by mistake were using an invalid field to check the condition.

https://linear.app/comm/issue/ENG-10111/local-dm-message-stalls-indefinitely

Test Plan: Created a pending message following https://linear.app/comm/issue/ENG-10111/local-dm-message-stalls-indefinitely#comment-fccc6bb4 scenario. Then applied this patch and noticed that the message was displayed correctly as failed and could be retried.

Reviewers: kamil, bartek

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14253