[CommRustModule][native_rust_library] rename `siwe*` variables to `siweSocialProof*`
Summary:
Part of [ENG-10112](https://linear.app/comm/issue/ENG-10112/generate-user-keys-backup-an-a-client-during-restoration) and [ENG-10113](https://linear.app/comm/issue/ENG-10113/send-the-user-keys-backup-to-identity-service-in-a-single-rpc)

We need to also pass SIWE message for backup now (which has to be uploaded together with User Keys), this could be confused with Social Proof message. This already caused some issues in the past when we mixed them so improving naming should help.

In the next diff I am going to update this on the JS side

Test Plan: Build the app

Reviewers: bartek, tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14255