[CommCoreModule] move `restoreUser` from `CommRustModule` to `CommCoreModule`
Summary:
Part of [ENG-10112](https://linear.app/comm/issue/ENG-10112/generate-user-keys-backup-an-a-client-during-restoration) and [ENG-10113](https://linear.app/comm/issue/ENG-10113/send-the-user-keys-backup-to-identity-service-in-a-single-rpc)

We need to add User Keys backup to restore RPC. To achieve that, we need to get Olm account from SQLite and Crypto Module to be able to pickle account. This couldn't be easily achieved from CommRustModule.

Alternative solutions:
1. Create User Keys backup in `CommCoreModule`, return it to JS and pass to restore RPC - bad because we can avoid passing bytes to JS.
2. Access SQLite and CryptoModule from `CommRustModule` - this is a lot more complex to implement and it is better to keep only one Turbo Module accessing DB.

Depends on D14256

Test Plan: Restore still works

Reviewers: bartek, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14257