[native_rust_library] create User Keys backup when restoring user
Summary:
[ENG-10113](https://linear.app/comm/issue/ENG-10113/send-the-user-keys-backup-to-identity-service-in-a-single-rpc)

Creating User Keys backup. Publishing this diff but it needs to be updated after the service part is finished.

NOTE: This stack should not be landed before the Identity is deployed on staging.

Depends on D14272, D14270

Test Plan: Test plan described [here](https://linear.app/comm/issue/ENG-10112/generate-user-keys-backup-an-a-client-during-restoration#comment-8cbdbd04)

Reviewers: bartek, tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14273