[lib] Set the restore flag to true for staff users
Summary:
When a staff user uses the web app, we should assume they have a migrated account. It means that we will use e.g. a new logout procedure for them.

https://linear.app/comm/issue/ENG-10127/logging-out-on-the-web-results-in-an-error

Test Plan: Created a new account on native (with the flag set to true), disconnected from Metro, opened the web app with a modified `isStaff` that always returns true, logged in and out and made sure that we used the new flow.

Reviewers: kamil, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14265