[native_rust_library][CommCoreModule] add User Keys params to restore user RPC
Summary:
Part of [ENG-10112](https://linear.app/comm/issue/ENG-10112/generate-user-keys-backup-an-a-client-during-restoration) and [ENG-10113](https://linear.app/comm/issue/ENG-10113/send-the-user-keys-backup-to-identity-service-in-a-single-rpc)

We need to add params needed to User Keys creation.

Not sure if we can somehow reduce the number of params here, but it started to reduce reabaillity.

Depends on D14271

Test Plan: Build app

Reviewers: bartek, tomek