[comm-lib] Add Send requirement for Defer closures
Summary:
This is to make `Defer` work in fully asynchronous context. It was working in actix-web without this because every HTTP listener is single-threaded there. For tonic gRPC, closures must be `Send` in RPC handlers in order to pass them between `await` calls

Depends on D14281

Test Plan: cargo check

Reviewers: kamil, tomek

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14282