[identity] Backup client method for preparing user keys
Summary:
Identity method to call Backup endpoint from D14276. It's similar to what we have in `backup_client` crate, but with error handling modified for Identity Service needs

Depends on D14276

Test Plan: Created unit test which calls this function and prints the output. Expected the same as in D14276 test plan

Reviewers: kamil, tomek

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14277