[native] Migrate recenter functionality in FullScreenViewModal
Summary:
https://linear.app/comm/issue/ENG-8150/convert-fullscreenviewmodal-to-reanimated-2-syntax

The recenter function triggers if there is no interaction (after some gesture ends and there is no animation). If the image is panned or zoomed out so there is some space left, the image is zoomed in to 1 (if necessary) and moved.

Depends on D14306

Test Plan: Tested after migration: verify the image is recentered after pan ended when zoomed out

Reviewers: tomek, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14307