[native] Migrate fling functionality in FullScreenViewModal
Summary:
https://linear.app/comm/issue/ENG-8150/convert-fullscreenviewmodal-to-reanimated-2-syntax

The fling function is triggered after pan ends. If the image is zoomed in and user moves the image with a pan gesture, the image moves a little with an inertia. Note that there recenter function will run instead if the image is moved to the edge of the screen.

The builds will fail due to some leftover reanimated v1 code. I'll remove it in the next diffs for diffs readability.

Depends on D14307

Test Plan: Tested after migration: there is an inertia effect after pan ended

Reviewers: tomek, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14308