[native] Remove leftover Reanimated v1 code in FullScreenViewModal
Summary:
https://linear.app/comm/issue/ENG-8150/convert-fullscreenviewmodal-to-reanimated-2-syntax

After migrating the last function from `updates` array, there is a lot of code left to be removed. For the readability of the diffs I removed it in one diff.
CI should pass now, the component still needs to be migrated completely to work.

Depends on D14309

Test Plan: flow

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14310