[native] Listen for a hide keyboard event on iOS
Summary:
This issue reproduces for me really infrequently, but this diff might be a good guess for a possible solution. According to the docs https://reactnative.dev/docs/keyboard#addlistener on Android only two event types are supported, but on iOS all the types are handled. Currently, we aren't listening for `keyboardDidHide` event on iOS, which looks like a possible source of the issue.

https://linear.app/comm/issue/ENG-10107/keyboard-area-is-not-dismissed-after-registration

Test Plan:
Tested that the chat changes its height after the keyboard is shown.

Reviewers: kamil, angelika